+++ src/lsp/constants.ts
@@ -4,10 +4,15 @@ import { posix } from 'path';
 
 export const SORBET_SCHEME = 'sorbet';
 
+export const SORBET_FILE_DOCUMENT_SELECTOR: vslc.DocumentFilter = {
+  language: 'ruby',
+  scheme: 'file',
+} as const;
+
 export const SORBET_DOCUMENT_SELECTOR: readonly vslc.DocumentFilter[] = [
-  { language: 'ruby', scheme: 'file' },
-  { language: 'ruby', scheme: SORBET_SCHEME },
-];
+  SORBET_FILE_DOCUMENT_SELECTOR,
+  { language: 'ruby', scheme: SORBET_SCHEME } as const,
+] as const;
 
 export function getWorkspaceDocumentSelector(
   workspaceFolder: vscode.WorkspaceFolder,