+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.1
  - **Run Ruby File** and **Debug Ruby File** commmands force-save files before running.
+ - Support navigation to `require_relative` target via a definition provider (enables **Peek**).
 
 ## 0.3.0
 - Multi-root workspace support, including per-folder Sorbet clients configurable individually or at the workspace level.